Add files via upload
This schematic is the one I am currently working on, in the simulation I am driving a parallel combination of a 47nF cap with 22uH inductor, their resonant frequency is 125kHz. I built and tested this circuit with TL740cn JFET op amps. they can handle this high frequency.

I'm currently working to get the output bigger than a few tens of mV with the load hooked up. I am going to simulate next instead of driving the inductor and cap, driving the output stage. I'm going to simulate with a class A output stage single BJT 4 resistor bias circuit. 

Please play around with feeding this filtered signal into different output driving stages.

I built the circuit using the pwm output as the input to the filter stage and it filters quite well. There are some harmonics in the sine wave, but it is really fairly sinusoidal. I will add pictures of the output waveform